Bump version to 2.0.0
