+++ CHANGELOG.md
@@ -5,4 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.1] - Unreleased
+## [2.0.0] - Unreleased
+
+### Removed
+
+- The `Circulator::Diverter` module has been and features added to `Circulator` itself.
