+++ lib/circulator/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Circulator
-  VERSION = "1.0.1"
+  VERSION = "2.0.0"
 end
